requirements.txt
flask
mysql_connector_python